Added the instantiate method to the Config object, as well as example useage.
