+++ skeletonkey/__init__.py
@@ -10,7 +10,8 @@ the dynamic loading of classes and their arguments at runtime.
 
 __version__ = "0.0.11"
 
-from .core import unlock, instantiate, instantiate_all, Config
+from .core import unlock, Config
+from .instantiate import instantiate, instantiate_all
 
 # Names to import with wildcard import
 __all__ = ["unlock", "instantiate", "instantiate_all", "Config"]
